Merges NFT Card Variants Into NftCard
